Convert script file into Addon Package
- Initial commit
- Converted script addon to addon package
