+++ src/r0fl_simple_toolbox/const.py
@@ -1,0 +1,11 @@
+from . import bl_info
+
+def _version_str(version_tuple: tuple):
+    v_str = '.'.join([str(n) for n in version_tuple])
+    print(f"{v_str=}")
+    return v_str
+
+VERSION = bl_info.get("version", (0, 0, 0))
+VERSION_STR = _version_str(VERSION)
+ADDON_NAME = bl_info.get("name")
+INTERNAL_NAME = "r0fl_simple_toolbox"
