+++ src/r0fl_simple_toolbox/utils.py
@@ -1,0 +1,210 @@
+import bpy
+import math
+
+from .const import INTERNAL_NAME
+
+def iter_scene_objects(selected=False, type: str = ''):
+        iters = bpy.data.objects
+        if selected:
+            iters = bpy.context.selected_objects
+            
+        for o in iters:
+            if not type or o.type == type:
+                yield o
+                
+def iter_children(p_obj, recursive=True):
+    """
+    Iterate through all children of a given parent object.
+    Args:
+        p_obj: Parent object to find children for
+        recursive: If True, also iterate through children of children
+    """
+    
+    for obj in bpy.data.objects:
+        if obj.parent == p_obj:
+            yield obj
+            if recursive:
+                yield from iter_children(obj, recursive=True)         
+
+def show_notification(message, title="Script Finished"):
+    """Display a popup notification and status info message"""
+    bpy.context.window_manager.popup_menu(lambda self, context: self.layout.label(text=message), title=title)
+    bpy.context.workspace.status_text_set(message)
+    
+def deselect_all():
+    bpy.ops.object.select_all(action="DESELECT")
+
+def save_preferences():
+    """Safely save user preferences without causing recursion"""
+    try:
+        if not hasattr(save_preferences, 'is_saving'):
+            save_preferences.is_saving = False
+            
+        if not save_preferences.is_saving:
+            save_preferences.is_saving = True
+            bpy.context.preferences.use_preferences_save = True
+            
+            bpy.data.scenes["Scene"].zen_uv.td_props.prp_current_td = get_td_value()
+            bpy.data.scenes["Scene"].zen_uv.td_props.td_unit = get_td_unit()
+            
+            bpy.ops.wm.save_userpref()
+            save_preferences.is_saving = False
+    except Exception as e:
+        print(f"Error saving preferences: {e}")
+        save_preferences.is_saving = False
+
+def get_td_value():
+    """Get the texel density value from addon preferences"""
+    try:
+        preferences = bpy.context.preferences.addons[INTERNAL_NAME].preferences
+        value = preferences.zenuv_td_prop
+        return value
+    except Exception as e:
+        print(e)
+        return 10.0  # default value if preferences not found
+    
+def get_td_unit():
+    """Get the texel density unit from addon preferences"""
+    try:
+        preferences = bpy.context.preferences.addons[INTERNAL_NAME].preferences
+        td_unit = preferences.zenuv_td_unit_prop
+        td_split = td_unit.split('_')
+        
+        if td_split and len(td_split) > 1:
+            td_unit = td_split[1].lower()
+        else:
+            td_unit = 'cm' # Default
+        
+        return td_unit
+    except Exception as e:
+        print(e)
+        return 'cm'  # default value if preferences not found
+
+def op_clear_sharp_along_axis(axis: str):
+    axis = str(axis).upper()
+    
+    threshold = bpy.context.preferences.addons[INTERNAL_NAME].preferences.clear_sharp_axis_float_prop
+    print(f"Threshold: {threshold}")
+    
+    # Collect select objects
+    objects = [obj for obj in bpy.context.selected_objects if obj.type == 'MESH']
+    
+    print(f"Objects: {objects}")
+    
+    if not objects:
+        return False
+    
+    for obj in objects:
+        # Set the active object
+        bpy.context.view_layer.objects.active = obj
+        print(f"Iterating: {obj.name}")
+        
+        # Check the mode
+        mode = obj.mode
+        print(f"Mode: {mode}")
+        
+        # Access mesh data
+        mesh = obj.data
+        print(f"Mesh: {mesh}")
+        
+        # Store the selection mode
+        # Tuple of Booleans for each of the 3 modes
+        selection_mode = tuple(bpy.context.scene.tool_settings.mesh_select_mode)
+        
+        # Store initial selections
+        # Vertices
+        selected_vertices = [v.index for v in mesh.vertices if v.select]
+        
+        # Edges
+        selected_edges = [e.index for e in mesh.edges if e.select]
+        
+        # Faces
+        selected_faces = [f.index for f in mesh.polygons if f.select]
+        
+        # Deselect all vertices
+        bpy.ops.object.mode_set(mode='EDIT')
+        bpy.ops.mesh.select_mode(type="VERT")
+        bpy.ops.mesh.select_all(action="DESELECT")
+        bpy.ops.object.mode_set(mode="OBJECT") # We're in Object mode so we can select stuff. Logic is weird.
+        
+        for idx, vertex in enumerate(mesh.vertices):
+            # print(f"Vertex {vertex.co}", end="")
+            
+            if axis == 'X':
+                if math.isclose(vertex.co.x, 0.0, abs_tol=threshold):
+                    mesh.vertices[idx].select = True
+                    # print(f" X isclose({vertex.co.x}, 0.0, abs_tol={threshold}): {math.isclose(vertex.co.x, 0.0, abs_tol=threshold)}")
+            
+            if axis == 'Y':
+                if math.isclose(vertex.co.y, 0.0, abs_tol=threshold):
+                    mesh.vertices[idx].select = True
+                    # print(f" Y isclose({vertex.co.y}, 0.0, abs_tol={threshold}): {math.isclose(vertex.co.y, 0.0, abs_tol=threshold)}")
+                    
+            if axis == 'Z':
+                if math.isclose(vertex.co.z, 0.0, abs_tol=threshold):
+                    mesh.vertices[idx].select = True
+                    # print(f" Z isclose({vertex.co.z}, 0.0, abs_tol={threshold}): {math.isclose(vertex.co.z, 0.0, abs_tol=threshold)}")
+        
+        # Enter Edit mode
+        bpy.ops.object.mode_set(mode="EDIT")
+        
+        # Switch to edge mode
+        bpy.ops.mesh.select_mode(use_extend=False, use_expand=False, type="EDGE")
+        
+        # Clear the Sharp
+        bpy.ops.mesh.mark_sharp(clear=True)
+        
+        # Restore the inital selections and mode
+        if selection_mode[0] is True:
+            # Deselect all vertices
+            bpy.ops.mesh.select_mode(type="VERT")
+            bpy.ops.mesh.select_all(action="DESELECT")
+            bpy.ops.object.mode_set(mode="OBJECT") # We're in Object mode so we can select stuff. Logic is weird.
+            for vert_idx in selected_vertices:
+                mesh.vertices[vert_idx].select = True
+        if selection_mode[1] is True:
+            # Deselect all vertices
+            bpy.ops.mesh.select_mode(type="EDGE")
+            bpy.ops.mesh.select_all(action="DESELECT")
+            bpy.ops.object.mode_set(mode="OBJECT") # We're in Object mode so we can select stuff. Logic is weird.
+            for edge_idx in selected_edges:
+                mesh.edges[edge_idx].select = True
+        if selection_mode[2] is True:
+            # Deselect all vertices
+            bpy.ops.mesh.select_mode(type="FACE")
+            bpy.ops.mesh.select_all(action="DESELECT")
+            bpy.ops.object.mode_set(mode="OBJECT") # We're in Object mode so we can select stuff. Logic is weird.
+            for face_idx in selected_faces:
+                mesh.polygons[face_idx].select = True
+        
+        # Set back to Object mode
+        bpy.ops.object.mode_set(mode=mode)
+
+
+def continuous_property_list_update(scene, context):
+    # This method is required to assess the last object selection, otherwise
+    # this is triggered on every click and the list is updated, and the checkboxes are reset
+    
+    if bpy.context.selected_objects:
+        current_selection = {obj.name for obj in iter_scene_objects(selected=True)}
+        addon_props = context.scene.r0fl_toolbox_props
+        prev_selection = set(addon_props.last_object_selection.split(',')) if addon_props.last_object_selection else set()
+
+        if current_selection != prev_selection:
+            addon_props.custom_property_list.clear()
+
+            # Add unique custom properties to the list
+            unique_props = set()
+            for obj in bpy.context.selected_objects:
+                for prop_name in obj.keys():
+                    if not prop_name.startswith('_') and prop_name not in unique_props:
+                        unique_props.add(prop_name)
+                        item = addon_props.custom_property_list.add()
+                        item.name = prop_name
+
+            # Update the last object selection
+            addon_props.last_object_selection = ','.join(current_selection)
+    else:
+        # Clear the property list if no objects are selected
+        context.scene.r0fl_toolbox_props.custom_property_list.clear()
+        context.scene.r0fl_toolbox_props.last_object_selection = ""
